sprint 3:hotfix: fix transaction getAll
