+++ NESTCOOKING_API.Presentation/Controllers/AdminController.cs
@@ -2,6 +2,8 @@
 using Microsoft.AspNetCore.Mvc;
 using NESTCOOKING_API.Business.DTOs;
 using NESTCOOKING_API.Business.DTOs.AdminDTOs;
+using NESTCOOKING_API.Business.DTOs.TransactionDTOs;
+using NESTCOOKING_API.Business.Services;
 using NESTCOOKING_API.Business.Services.IServices;
 using static NESTCOOKING_API.Utility.StaticDetails;
 
@@ -15,11 +17,12 @@ namespace NESTCOOKING_API.Presentation.Controllers
     {
         private readonly IResponseService _responseService;
         private readonly IReportService _reportService;
-
-        public AdminController(IResponseService responseService, IReportService reportService)
+        private readonly ITransactionService _transactionService;
+        public AdminController(IResponseService responseService, IReportService reportService,ITransactionService transactionService)
         {
             _responseService = responseService;
             _reportService = reportService;
+            _transactionService = transactionService;
         }
         [HttpPost("reports/{reportId}")]
         public async Task<IActionResult> HandleReport([FromBody] AdminRequestDTO adminRequestDTO)
@@ -54,6 +57,23 @@ namespace NESTCOOKING_API.Presentation.Controllers
                 return BadRequest(ResponseDTO.Create(System.Net.HttpStatusCode.BadRequest, ex.Message));
             }
         }
-           
+        [HttpGet("transaction")]
+        public async Task<ActionResult<List<TransactionDTO>>> GetAllTransactions()
+        {
+            try
+            {
+                var result = await _transactionService.GetAllTransactions();
+                if (result == null)
+                {
+                    return BadRequest(ResponseDTO.BadRequest(message: "Not found transaction by user"));
+                }
+                return Ok(ResponseDTO.Accept(result: result));
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ResponseDTO.BadRequest(message: ex.Message));
+            }
+        }
+
     }
 }
